+++ src/games/fallout4vr/fallout4vrsavegame.h
@@ -18,8 +18,8 @@ public:
 
 protected:
   // Fetch easy-to-access information.
-  void fetchInformationFields(FileWrapper& file, unsigned long& saveNumber,
-                              QString& playerName, unsigned short& playerLevel,
+  void fetchInformationFields(FileWrapper& file, uint32_t& saveNumber,
+                              QString& playerName, uint16_t& playerLevel,
                               QString& playerLocation, FILETIME& creationTime) const;
 
   std::unique_ptr<DataFields> fetchDataFields() const override;
