+++ src/games/starfield/starfieldsavegame.h
@@ -21,9 +21,9 @@ protected:
   // Fetch easy-to-access information.
   void getData(FileWrapper& file) const;
 
-  void fetchInformationFields(FileWrapper& file, unsigned long& saveNumber,
+  void fetchInformationFields(FileWrapper& file, uint32_t& saveNumber,
                               unsigned char& saveVersion, QString& playerName,
-                              unsigned short& playerLevel, QString& playerLocation,
+                              uint16_t& playerLevel, QString& playerLocation,
                               FILETIME& creationTime) const;
 
   std::unique_ptr<DataFields> fetchDataFields() const override;
